Varredura completa de bugs (crítico a baixo) + anamnese inicial completa
Cobre a varredura ponta a ponta pedida (segurança de upload, race conditions,
validações ausentes, notificações vazando dado sensível, rate limiting, etc.)
e complementa a anamnese inicial do aluno com as seções do formulário em papel
do personal que ainda não existiam no app (histórico de atividade física,
objetivos detalhados, condições de saúde, estilo de vida, motivação/
preferências, disponibilidade, histórico familiar) — removendo em seguida o
card de PAR-Q editável da ficha do aluno, que ficou redundante com a anamnese
completa. Cada correção tem teste cobrindo o caso que ela resolve.

+++ backend-laravel/app/Http/Controllers/NegocioController.php
@@ -49,7 +49,10 @@ class NegocioController extends Controller
         $diasSemTreinar = self::DIAS_SEM_TREINAR;
         $diasSemCheckin = self::DIAS_SEM_CHECKIN;
         $diasAlunoNovo = self::DIAS_ALUNO_NOVO;
-        $treinosMinimosAlunoNovo = self::TREINOS_MINIMOS_ALUNO_NOVO;
+        // Interpolada direto num CASE da query (constante da classe, nunca vem de
+        // input) — cast explícito por defesa em profundidade, mesmo padrão de
+        // Estagnacao::compararUltimasSessoes.
+        $treinosMinimosAlunoNovo = (int) self::TREINOS_MINIMOS_ALUNO_NOVO;
 
         // date_trunc/interval não existem no MySQL — limites de data calculados em PHP/Carbon.
         $agora = Carbon::now()->toDateTimeString();
